Add favicon. Create context panel and style it. Extract some styles to mixins.js. Move global, styles related files to separate folder 'styles'
