+++ client/src/components/App/App.js
@@ -1,15 +1,24 @@
-import styled from "styled-components";
+import styled from 'styled-components';
+import {baseShadow, absoluteCenter} from '../../styles/mixins';
 
 function App() {
   return (
-    <div>
+    <Wrapper>
       Hello world
-    </div>
+    </Wrapper>
   );
 }
 
 const Wrapper = styled.div`
-  
+    width: min(90%, 1800px);
+    height: min(90%, 950px);
+    ${absoluteCenter}
+    
+    background-color: var(--color-background);
+    ${baseShadow}
+
+    border-radius: 20px;
+    padding: 58px 56px;
 `;
 
 export default App;
