--- client/src/constants.js
@@ -1,6 +1,0 @@
-const COLORS = {
-    primary: '#F79D75',
-    secondary: '#6A7588',
-    grey: '#BFC0C0',
-    background: '#BFC0C0'
-}
