+++ client/src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './components/App';
 import {BrowserRouter, Routes, Route} from 'react-router-dom';
-import GlobalStyles from './globalStyles';
+import GlobalStyles from './styles/globalStyles';
 import AboutPage from './pages/AboutPage';
 import HomePage from './pages/HomePage';
