+++ client/src/styles/mixins.js
@@ -1,0 +1,12 @@
+import {css} from 'styled-components';
+
+export const baseShadow = css`
+    box-shadow: 0 0 17px rgba(0,0,0,.15);
+`;
+
+export const absoluteCenter = css`
+    position: absolute;
+    left: 50%;
+    top: 50%;
+    transform: translate(-50%, -50%);
+`;
